+++ CHANGELOG.md
@@ -3,6 +3,7 @@ English | [简体中文](./CHANGELOG_CN.md)
 ## 3.14.4-rc (2022-03-??)
 
 - `Fix(Network)` Fix CPU high load bug when response is a large string. (issue #515)
+- `Fix(Network)` Fix missing Request Headers issue in XHR. (issue #522)
 
 
 ## 3.14.3 (2022-03-28)